Restructured demo folder. Added authorship and comments. Cleaned up unnecessary code.
