+++ demo/compile.command
@@ -1,6 +1,17 @@
+<<com
+/**
+ * @file compile.command
+ * @author Robert Oganesian (roganesian@uri.edu)
+ * @brief macOS script to compile the y-fast-trie GUI.
+ * @version 1.0
+ * @date 2022-04-25
+ * 
+ */
+com
+
 #! /bin/bash
 
-OBJS="main.cpp visualizer.cpp console-menu.cpp states/state.cpp states/visualizer-state.cpp states/main-menu-state.cpp states/about-state.cpp tools/gui.cpp"
+OBJS="main.cpp"
 INC="-I/usr/local/Cellar/sfml/2.5.1_1/include"
 LIB="-L/usr/local/Cellar/sfml/2.5.1_1/lib -lsfml-graphics -lsfml-window -lsfml-system -lsfml-audio"
