--- demo/config/about-state-keys.ini
@@ -1,1 +1,0 @@
-CLOSE Escape
