+++ demo/resource/image/visualizer.dot
@@ -1,0 +1,75 @@
+digraph {
+	node[style=filled, fillcolor=white];
+	n0x600002736560 -> n0x6000027363e0
+	n0x600002736560 -> n0x60000273b580
+	n0x600002736560[label="0"]
+	{rank = same; n0x600002736560; }
+	n0x60000273b580 -> n0x60000273b920
+	n0x60000273b580 -> n0x60000274d580
+	n0x60000273b580[label="1"]
+	n0x6000027363e0 -> n0x600002736480
+	n0x6000027363e0 -> n0x6000027363a0
+	n0x6000027363e0[label="0"]
+	{rank = same; n0x60000273b580; n0x6000027363e0; }
+	n0x60000274d580 -> n0x60000273b920
+	n0x60000274d580 -> n0x60000274c7c0
+	n0x60000274d580[label="3"]
+	n0x6000027363a0 -> n0x600002736480
+	n0x6000027363a0 -> n0x6000027365a0
+	n0x6000027363a0[label="1"]
+	{rank = same; n0x60000274d580; n0x6000027363a0; }
+	n0x60000274c7c0 -> n0x60000273b920
+	n0x60000274c7c0 -> n0x60000274d3c0
+	n0x60000274c7c0[label="7"]
+	n0x6000027365a0 -> n0x600002736480
+	n0x6000027365a0 -> n0x6000027365c0
+	n0x6000027365a0[label="3"]
+	{rank = same; n0x60000274c7c0; n0x6000027365a0; }
+	n0x60000274d3c0 -> n0x60000273b920
+	n0x60000274d3c0 -> n0x60000274c7e0
+	n0x60000274d3c0[label="15"]
+	n0x6000027365c0 -> n0x600002736480
+	n0x6000027365c0 -> n0x6000027364c0
+	n0x6000027365c0[label="7"]
+	{rank = same; n0x60000274d3c0; n0x6000027365c0; }
+	n0x60000274c7e0 -> n0x60000273b920
+	n0x60000274c7e0 -> n0x60000274d420
+	n0x60000274c7e0[label="31"]
+	n0x6000027364c0 -> n0x600002736480
+	n0x6000027364c0 -> n0x6000027364a0
+	n0x6000027364c0[label="15"]
+	{rank = same; n0x60000274c7e0; n0x6000027364c0; }
+	n0x60000274d420 -> n0x60000273b920
+	n0x60000274d420 -> n0x6000027357a0
+	n0x60000274d420[label="63"]
+	n0x6000027364a0 -> n0x600002736480
+	n0x6000027364a0 -> n0x600002736460
+	n0x6000027364a0[label="31"]
+	{rank = same; n0x60000274d420; n0x6000027364a0; }
+	n0x6000027357a0 -> n0x60000273b920
+	n0x6000027357a0 -> n0x60000273b920
+	n0x6000027357a0[label="127"]
+	n0x600002736460 -> n0x600002736480
+	n0x600002736460 -> n0x600002736480
+	n0x600002736460[label="63"]
+	{rank = same; n0x6000027357a0; n0x600002736460; }
+	n0x60000273b920 -> n0x600002736480
+	n0x60000273b920[label="255"]
+	n0x600002736480 -> n0x60000273b920
+	n0x600002736480[label="127"]
+	{rank = same; n0x60000273b920; n0x600002736480; }
+n0x60000273b920 -> n0x60000274d4c0;
+subgraph RBTree0x60000274d4c0 {
+
+	n0x60000274d4c0 [label="255", color=black, fillcolor="#bebebe", style=filled];
+
+}
+n0x600002736480 -> n0x600002720c00;
+subgraph RBTree0x600002720c00 {
+
+	n0x600002720c00 [label="123", color=black, fillcolor="#bebebe", style=filled];
+	n0x60000272c1a0 [label="113", color=red, fillcolor="#ff6a6a", style=filled];
+	n0x600002720c00 -> {n0x60000272c1a0};
+
+}
+}
