--- demo/visualizer.h
@@ -1,52 +1,0 @@
-#pragma once
-
-#include "states/main-menu-state.h"
-#include "states/visualizer-state.h"
-
-class Visualizer {
-private:
-    // Variables
-    sf::RenderWindow *window;
-    sf::Event event;
-    std::vector<sf::VideoMode> videoModes;
-    sf::ContextSettings window_settings;
-
-    sf::Clock dt_clock;
-    float dt;
-
-    // Abstract class cannot be created as an object
-    std::stack<State*> states;
-
-    std::map<std::string, int> valid_keys;
-
-    // Initializers
-    void initWindow();
-    void initKeys();
-    void initStates();
-
-public:
-    // Constructors/Destructors
-    Visualizer();
-    ~Visualizer();
-
-    bool isVisualizerState();
-    void updateVisualizerBackground();
-
-    // Functions //
-
-    // Update
-    void updateDt();
-    void updateEvents();
-    void update();
-    
-    // Render
-    void render();
-    
-    // Core
-    void run();
-    void endApplication();
-
-};
-
-extern Visualizer visualizer;
-extern void UPDATE_GUI();
